[WIP] Address feedback on moving embedding-service errors to error module (#20)
* Initial plan

* refactor: move embedding service errors to application layer


---------
