+++ apps/api-cli/src/application/errors/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Application Errors barrel export
+ */
+
+export {
+  EmbeddingServiceError,
+  EmbeddingServiceUnavailableError,
+} from './ApplicationErrors.js';
