+++ apps/api-cli/src/application/index.ts
@@ -3,3 +3,4 @@
  */
 
 export * from './ports/index.js';
+export * from './errors/index.js';
